+++ ChangeLog.md
@@ -89,6 +89,9 @@
 
   * Add full grading job log display from S3 (Nathan Walters).
 
+  * Add instructor editing of total points and question points for assessment
+    instances (Matt West).
+
   * Split installing documentation into separate method sections (Matt West).
 
   * Remove unused dead code (`/lib/db.js`, `question-servers/shortAnswer.js`,